some quick updates
landclaim now has permission checks, as well as overall availibleland
and claimed land amount checks.

/shop was added for survival, only current item is to buy more land. Set
to 10 tickets for the beta period, easy to change to real cost

::Still need to add cost per place, with a free place for the first
chunk::

nothing that tricky, but that may have to wait for launch, instead of
people wasting to many tickets on land they will lose.