Create TicketLogSystem
A little subsystem that all ticket transactions will refer to for logging. Will make it easier for life as we know it as staff members to keep track of things later down the line. Hopefully Jack will be able to easily parse into tables for the text log files he wants.